+++ Runtime/Extensions/CollectionExtensions.cs
@@ -14,6 +14,20 @@ namespace CinderUtils.Extensions {
             if (collection.Count == 0) return default(T);
             else return collection.ElementAt(Random.Range(0, collection.Count));
         }
+
+        public static T LoopingGet<T>(this ICollection<T> collection, int i) {
+            if (collection.Count == 0) return default(T);
+            else return collection.ElementAt(i % collection.Count);
+        }
+    }
+
+    // Extensions for Collections
+    public static class DictionaryExtensions {
+
+        public static V GetValue<K, V>(this IDictionary<K, V> dict, K key, V defaultValue = default(V)) {
+            return dict.TryGetValue(key, out var value) ? value : defaultValue;
+        }
     }
 
 }
+